Add files via upload
Why adjust for the size of the population?

Differences in the population size between countries are often large, and the COVID-19 death count in more populous countries tends to be higher. Because of this it can be insightful to know how the number of confirmed deaths in a country compares to the number of people who live there, especially when comparing across countries.

For instance, if 1,000 people died in Iceland, out of a population of about 340,000, that would have a far bigger impact than the same number dying in the United States, with its population of 331 million.1 This difference in impact is clear when comparing deaths per million people of each country’s population – in this example it would be roughly 3 deaths/million people in the US compared to a staggering 2,941 deaths/million people in Iceland.